+++ src/cli/sections/InteractionSection.py
@@ -1,0 +1,14 @@
+from rich.prompt import Prompt
+
+
+# class Interaction:
+#     def __init__(self) -> None:
+#         self._question: str = ""
+
+#     def render(self):
+#         return Prompt(self._question)
+
+#     def set_question(self, question: str):
+#         self._question = question
+
+# class IN