+++ src/config/types.py
@@ -1,37 +1,55 @@
 """All game custom types"""
-from typing import Literal
+# from ..Card import Card
+from typing import Annotated, Literal, Type, TypeAlias, TypedDict, Any, TypeGuard
 
-CardDataType = dict[Literal["id"] : int, Literal["value"] : str, Literal["suit"] : str]
-
-PlayerDataType = dict[Literal["id"] : int, Literal["hand"] : list]
-PlayerHandType = list[CardDataType]
-
-CardSuitsType = Literal["spades", "clubs", "hearts", "diamonds", "metacard", "joker"]
-CardValuesType = Literal[
-    "1",
-    "2",
-    "3",
-    "4",
-    "5",
-    "6",
-    "7",
-    "8",
-    "9",
-    "10",
-    "11",
-    "12",
-    "13",
-    "14",
-    "15",
-    "16",
-    "17",
-    "18",
-    "19",
-    "20",
-    "21",
-    "V",
-    "C",
-    "Q",
-    "K",
-    "JOKER",
+CardSuitsType: TypeAlias = Literal[
+    "spades", "clubs", "hearts", "diamonds", "metacard", "joker"
 ]
+
+# str here cause of metacard values
+CardValuesType: TypeAlias = (
+    str
+    | Literal[
+        "1",
+        "2",
+        "3",
+        "4",
+        "5",
+        "6",
+        "7",
+        "8",
+        "9",
+        "10",
+        "11",
+        "12",
+        "13",
+        "14",
+        "15",
+        "16",
+        "17",
+        "18",
+        "19",
+        "20",
+        "21",
+        "V",
+        "C",
+        "Q",
+        "K",
+        "JOKER",
+    ]
+)
+# CardDataType:TypeAlias = dict[Literal["id","value", "suit",], str | int]
+class CardDataType(TypedDict):
+    id: int
+    value: CardValuesType
+    suit: CardSuitsType
+
+
+# -- PLAYER TYPES
+# def is_card_list(val: list[Card]) -> TypeGuard[list[Card]]:
+
+
+# PlayerDataType:TypeAlias = dict[Literal["id","hand"], int| list]
+class PlayerDataType(TypedDict):
+    id: int
+    hand: list